LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/sticky/view-timeline-sticky-offscreen-7-expected.txt

Subject

FAIL View timeline target > viewport, bottom-sticky and top-sticky during contain. assert_equals: Effect at the start of the active phase: cover 0% to cover 100% expected "0.3" but got "0.336364"
